backend2/requirements.txt
fastapi
uvicorn
motor
passlib[bcrypt]
bcrypt>=4.0.0
python-jose[cryptography]
python-multipart
email-validator
python-dotenv
pymongo
httpx
PyPDF2
jinja2
python-pptx
yt-dlp
exa-py
apscheduler
python-jobspy
python-dateutil
openai
PyMuPDF
aiohttp